Keep different classes in different files
