+++ src/main/scala/io/github/otbproject/otb/plugin/content/PluginDataTypeIdentifier.scala
@@ -1,0 +1,3 @@
+package io.github.otbproject.otb.plugin.content
+
+private[content] final case class PluginDataTypeIdentifier[T <: PluginData](plugin: ContentPlugin, tClass: Class[T])
