+++ src/main/scala/io/github/otbproject/otb/plugin/content/data/Data.scala
@@ -34,8 +34,3 @@ private[data] abstract class Data[T](provider: PluginDataFactory => DataFactory[
     builder.build
   }
 }
-
-final class RepeatedInitializationException private[content](data: Data[_])
-  extends IllegalStateException("Attempted to generate PluginData for [" + data + "] more than once. " +
-                                  "This was probably caused by some plugin attempting to access data from the object " +
-                                  "during its initialization.")
